requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via aiogram
aiogram>=3.20.0
    # via -r requirements.in
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via aiogram
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
certifi==2025.4.26
    # via
    #   aiogram
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   yarl
iniconfig==2.1.0
    # via pytest
jiter==0.9.0
    # via openai
levenshtein==0.27.1
    # via python-levenshtein
magic-filter==1.0.12
    # via aiogram
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
openai>=1.77.0,<2.0.0
    # via -r requirements.in
packaging==25.0
    # via pytest
pluggy==1.5.0
    # via pytest
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r requirements.in
pydantic==2.11.4
    # via
    #   aiogram
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements.in
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-vcr
pytest-asyncio==0.26.0
    # via -r requirements.in
pytest-vcr==1.0.2
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   pydantic-settings
python-levenshtein==0.27.1
    # via -r requirements.in
pyyaml==6.0.2
    # via vcrpy
rapidfuzz==2.13.7
    # via
    #   -r requirements.in
    #   levenshtein
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tomli==2.2.1
    # via pytest
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   aiogram
    #   anyio
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
urllib3==1.26.20
    # via vcrpy
vcrpy==7.0.0
    # via pytest-vcr
wrapt==1.17.2
    # via vcrpy
yarl==1.20.0
    # via
    #   aiohttp
    #   vcrpy

# Added for image processing
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Added for caching
redis>=6.0.0

# Added for HTTP requests
requests>=2.30.0
